make env and utils and rerun notebooks
