+++ 0.wave1_data/1.cellprofiler_ic_processing/cellprofiler_environment.yml
@@ -20,5 +20,5 @@ dependencies:
   - wxpython=4.1.0
   - mysqlclient=1.4.4
   - sentry-sdk=0.18.0
-- pip:
-    - cellprofiler==4.2.4
+  - pip:
+      - cellprofiler==4.2.4
