--- 0.wave1_data/1.cellprofiler_ic_processing/create_loaddata_csv.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-jupyter nbconvert --to python create_loaddata_csv.ipynb
-python create_loaddata_csv.py
