+++ 0.wave1_data/1.cellprofiler_ic_processing/create_loaddata_wave1_csv.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# # Create LoadData with illumination correction functions csv for CellProfiler
+# # Create LoadData csv for CellProfiler `illum.cppipe` pipeline - Wave 1
 
 # ## Import libraries
 
@@ -26,15 +26,12 @@ import loaddata_utils as ld
 index_directory = pathlib.Path("/home/jenna/Interstellar_Project/0.wave1_data/0.download_data/70117_20230118MM1_AbTest_V2__2023-01-25T16_44_54-Measurement1/Images")
 config_path = pathlib.Path("/home/jenna/Interstellar_Project/0.wave1_data/1.cellprofiler_ic_processing/config.yml")
 path_to_output = pathlib.Path("/home/jenna/Interstellar_Project/0.wave1_data/1.cellprofiler_ic_processing/wave1_loaddata.csv")
-illum_directory = pathlib.Path("/home/jenna/Interstellar_Project/0.wave1_data/1.cellprofiler_ic_processing/illum_directory")
-plate_id = "70117_20230118MM1_AbTest_V2"
-illum_output_path = pathlib.Path("/home/jenna/Interstellar_Project/0.wave1_data/2.cellprofiler_analysis/wave1_loaddata_with_illum.csv")
 
 
-# ## Create the LoadData csv with illum correction functions
+# ## Create the LoadData csv
 
 # In[3]:
 
 
-ld.create_loaddata_illum_csv(index_directory, config_path, path_to_output, illum_directory, plate_id, illum_output_path)
+ld.create_loaddata_csv(index_directory, config_path, path_to_output)
