+++ 0.wave1_data/1.cellprofiler_ic_processing/create_loaddata_wave1_csv.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+jupyter nbconvert --to python create_loaddata_wave1_csv.ipynb
+python create_loaddata_wave1_csv.py
